+++ hutool-cron/src/main/java/cn/hutool/cron/CronTimer.java
@@ -37,7 +37,7 @@ public class CronTimer extends Thread implements Serializable {
 	
 	@Override
 	public void run() {
-		final long timerUnit = this.scheduler.matchSecond ? TIMER_UNIT_SECOND : TIMER_UNIT_MINUTE;
+		final long timerUnit = this.scheduler.config.matchSecond ? TIMER_UNIT_SECOND : TIMER_UNIT_MINUTE;
 		
 		long thisTime = System.currentTimeMillis();
 		long nextTime;